pathname: carefully refactor finishing(?) touches
Return NIL as opposed to signalling error with PATHNAME-JAR-P and
PATHNAME-URL-P.
