+++ src/org/armedbear/lisp/Site.java
@@ -57,7 +57,7 @@ public final class Site
             if (!Pathname.isSupportedProtocol(url.getProtocol())) {
                 LISP_HOME = NIL;
             } else {
-              Pathname p = (Pathname)Pathname.create(url);
+              Pathname p = (Pathname)PathnameURL.create(url);
               p.setName(NIL).setType(NIL);
               LISP_HOME = p;
             }
